requirements.txt
blessings==1.6
curtsies==0.2.6
Django==1.9.7
django-pagedown==0.1.1
djangorestframework==3.4.5
greenlet==0.4.10
mccabe==0.5.2
Pillow==3.3.0
psycopg2==2.6.2
pycodestyle==2.0.0
pyflakes==1.2.3
Pygments==2.1.3
requests==2.11.0
six==1.10.0
wcwidth==0.1.7
